minor graph middleware refactor, no logic changes (#3360)
